1. retention policy for trash paths to 1 day

+++ src/main/java/com/inmobi/databus/purge/DataPurgerService.java
@@ -113,13 +113,11 @@ public class DataPurgerService extends AbstractService {
   }
 
   private Integer getDefaultStreamRetentionInDays() {
-    // retentionperiod is 1 day, setting to 2 to avoid overlap
-    return new Integer(2);
+    return new Integer(1);
   }
 
   private Integer getTrashPathRetentionInDays() {
-    // retentionperiod is 1 day, setting to 2 to avoid overlap
-    return new Integer(2);
+    return new Integer(1);
   }
 
   private Integer getRetentionPeriod(String streamName) {
